UBB-CS/Year2/Semester 2/Medii de proiectare si programare/Labs/Lab5/Backend/requirements.txt
Django==4.1.7
django_filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
django-cors-headers==3.14.0
pytz==2022.7.1
drf_yasg